% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{clean_data}
\alias{clean_data}
\title{clean_data}
\usage{
clean_data(
  csv,
  emissions_conversions = NULL,
  by_month = month,
  by_enduse = enduse,
  by_hour = Hour,
  by_fuel = fuel,
  ...
)
}
\arguments{
\item{csv}{character string of path to csv}

\item{emissions_conversions}{character string of path to csv with tonne CO2-e/MWh and tonne CO2-e/therm conversion factors}

\item{by_month}{designates whether to summarize data by month (the default) or omit month (NULL)}

\item{by_enduse}{designates whether to summarize data by enduse (the default) or omit enduse (NULL)}

\item{by_hour}{designates whether to summarize data by hour (the default) or omit hour (NULL)}

\item{by_fuel}{designates whether to summarize data by fuel (the default) or omit fuel (NULL)}

\item{...}{passes arguments to convert()}
}
\value{
Cleaned dataframe
}
\description{
clean_data
}
