% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{clean_enduse}
\alias{clean_enduse}
\title{clean_enduse}
\usage{
clean_enduse(data)
}
\arguments{
\item{data}{a dataframe with an enduse column (column name must match exactly). Column type must be chr.}
}
\value{
dataframe with cleaned enduse column
}
\description{
clean_enduse
}
