% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{convert}
\alias{convert}
\title{convert}
\usage{
convert(num, conversion_factor = 3600000)
}
\arguments{
\item{num}{a numeric representing energy use}

\item{conversion_factor}{a numeric representing the conversion factor from num's units to kWh. Default is 3600000. 0 throws an error.}
}
\value{
Energy in kWh
}
\description{
convert
}
