% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplemental_plotting_functions.R
\name{plot_colors}
\alias{plot_colors}
\title{plot_colors}
\usage{
plot_colors(data)
}
\arguments{
\item{data}{dataframe containing enduse column of type chr.}
}
\value{
Vector of hex codes
}
\description{
plot_colors
}
