% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_comps}
\alias{plot_comps}
\title{plot_comps}
\usage{
plot_comps(
  baseline,
  proposed,
  title = NULL,
  bw = FALSE,
  result = "Consumption",
  ...
)
}
\arguments{
\item{baseline}{character string of path to csv containing baseline data}

\item{proposed}{character string of path to csv containing proposed data}

\item{title}{character string of desired plot title. Default is NULL}

\item{bw}{boolean designating whether to plot in color (FALSE, default), or black and white (TRUE)}

\item{result}{"Consumption" or "Emissions" - denotes whether plot displays energy consumption (in kWh) or CO2 emissions (in grams CO2)}

\item{...}{passes arguments to clean_data()}
}
\value{
line plot comparing two models
}
\description{
plot_comps
}
