% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_dualfuel_avgs}
\alias{plot_dualfuel_avgs}
\title{plot_dualfuel_avgs}
\usage{
plot_dualfuel_avgs(csv, title = NULL, ...)
}
\arguments{
\item{csv}{character string of path to csv}

\item{title}{character string of desired plot title. Default is NULL}

\item{...}{passes arguments to clean_data()}
}
\value{
line plot of dual-fuel models
}
\description{
plot_dualfuel_avgs
}
