% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_emissions}
\alias{plot_emissions}
\title{plot_emissions}
\usage{
plot_emissions(model, ...)
}
\arguments{
\item{model}{character string of path to csv containing model data}

\item{...}{passes arguments to plot_model()}

\item{emissions_conversions}{character string of path to csv with tonne CO2-e/MWh and tonne CO2-e/therm conversion factors}
}
\value{
line plot displaying emissions of one model
}
\description{
plot_emissions
}
