% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplemental_plotting_functions.R
\name{plot_linetypes}
\alias{plot_linetypes}
\title{plot_linetypes}
\usage{
plot_linetypes(data, plot_aes)
}
\arguments{
\item{data}{dataset with "fuel" column. Column must be character or factor, and have either "Electricity" or "Gas"}

\item{plot_aes}{designates whether output is line type (line) or pattern (pattern)}
}
\value{
Vector of linetypes or patterns
}
\description{
plot_linetypes
}
