% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_model}
\alias{plot_model}
\title{plot_model}
\usage{
plot_model(model, title = NULL, result = "Consumption", ...)
}
\arguments{
\item{model}{character string of path to csv containing model data}

\item{title}{character string of desired plot title. Default is NULL}

\item{result}{"Consumption" or "Emissions" - denotes whether plot displays energy consumption (in kWh) or CO2 emissions (in grams CO2)}

\item{...}{passes arguments to clean_data()}
}
\value{
line plot displaying emissions or energy consumption of one model
}
\description{
plot_model
}
