% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_stacked_enduses}
\alias{plot_stacked_enduses}
\title{plot_stacked_enduses}
\usage{
plot_stacked_enduses(
  baseline,
  proposed = NULL,
  title = NULL,
  by_fuel = fuel,
  result = "Consumption",
  ...
)
}
\arguments{
\item{baseline}{character string of path to csv containing data}

\item{proposed}{character string of path to csv containing data to compare, or NULL (default)}

\item{title}{character string of desired plot title. Default is NULL}

\item{by_fuel}{designates whether to summarize data by fuel (the default) or omit fuel (NULL)}

\item{result}{"Consumption" or "Emissions" - denotes whether plot displays energy consumption (in kWh) or CO2 emissions (in grams CO2)}

\item{...}{passes arguments to clean_data()}
}
\value{
Stacked barchart of enduses
}
\description{
plot_stacked_enduses
}
