% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{to_remove}
\alias{to_remove}
\title{to_remove}
\usage{
to_remove(data)
}
\arguments{
\item{data}{a dataframe with a fuel column (column name must match exactly). Column type must be chr.}
}
\value{
String with fuel to remove or NULL
}
\description{
to_remove
}
