% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_enduse_avgs}
\alias{plot_enduse_avgs}
\title{plot_enduse_avgs}
\usage{
plot_enduse_avgs(csv, title = NULL, bw = FALSE, result = "Consumption", ...)
}
\arguments{
\item{csv}{character string of path to csv}

\item{title}{character string of desired plot title. Default is NULL}

\item{bw}{boolean designating whether to plot in color (FALSE, default), or black and white (TRUE)}

\item{result}{"Consumption" or "Emissions" - denotes whether plot displays energy consumption (in kWh) or CO2 emissions (in grams CO2)}

\item{...}{passes arguments to clean_data()}
}
\value{
line plot with enduse averages
}
\description{
plot_enduse_avgs
}
